Adding todo logic
