--- src/components/InputForm.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-
-
-export default Header
