--- src/components/ToDoItem.js
@@ -1,4 +1,0 @@
-// this is a single ToDo item
-import React, { Component } from 'react';
-
-export default ToDoItem
