+++ src/components/ToDoItemComponents/ToDoItem.css
@@ -1,0 +1,10 @@
+.ToDoItem{
+  /* STYLES*/
+  
+  
+  /* DEV SETTINGS */
+   border: solid;
+   border-width: .75px;
+   border-color: rgb(165, 89, 197);
+  }
+
