+++ src/components/ToDoItemComponents/ToDoItem.js
@@ -1,0 +1,20 @@
+// this is a single ToDo item
+import React, { Component } from 'react';
+
+import './ToDoItem.css';
+
+const ToDoItem = () => {
+  return(
+    <div>
+      <ul>
+        <li>
+          todo item is here
+        </li>
+
+      </ul>
+
+    </div>
+  )
+}
+
+export default ToDoItem
