--- src/components/ToDoList.js
@@ -1,15 +1,0 @@
-// this is the list of ToDo items
-import React, { Component } from 'react';
-import './ToDoList.css';
-
-const ToDoList = () => {
-  return(
-    <div>
-      <div>
-        <h2>Todo App</h2>
-      </div>
-      <ToDoItem>ToDo item list appears here.</ToDoItem>
-
-    </div>
-  )
-export default ToDoList;
